requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
numpy
python-dotenv
